# services/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API Dashboard BPS (master \u0026 insight endpoints) - schema webapi.",
        "title": "Dashboard BPS API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/datacontent": {
            "get": {
                "description": "Get list datacontent (payload data)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master"
                ],
                "summary": "List datacontent",
                "parameters": [
                    {
                        "type": "string",
                        "example": "[{\"field\":\"domain_id\",\"operator\":\"eq\",\"value\":\"0000\"}]",
                        "description": "JSON array of {field,operator,value}. Operator supports: eq, neq, gt, gte, lt, lte, like",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "datacontent_value",
                        "description": "sort column (single)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "asc | desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "per page, default 20",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "page, 1-based",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.DataContentListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/master/domain": {
            "get": {
                "description": "Get list master domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master"
                ],
                "summary": "List master domain",
                "parameters": [
                    {
                        "type": "string",
                        "example": "[{\"field\":\"domain_id\",\"operator\":\"eq\",\"value\":\"0000\"}]",
                        "description": "JSON array of {field,operator,value}. Operator supports: eq, neq, gt, gte, lt, lte, like",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "domain_name",
                        "description": "sort column (single)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "asc | desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "per page, default 20",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "page, 1-based",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.DomainListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    }
                }
            }
        },
        "/master/subjek": {
            "get": {
                "description": "Get list master subject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master"
                ],
                "summary": "List master subject",
                "parameters": [
                    {
                        "type": "string",
                        "example": "[{\"field\":\"domain_id\",\"operator\":\"eq\",\"value\":\"0000\"}]",
                        "description": "JSON array of {field,operator,value}. Operator supports: eq, neq, gt, gte, lt, lte, like",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "sub_name",
                        "description": "sort column (single)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "asc | desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "per page, default 20",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "page, 1-based",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.SubjectListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    }
                }
            }
        },
        "/master/tahun": {
            "get": {
                "description": "Get list master tahun",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master"
                ],
                "summary": "List master tahun",
                "parameters": [
                    {
                        "type": "string",
                        "example": "[{\"field\":\"domain_id\",\"operator\":\"eq\",\"value\":\"0000\"}]",
                        "description": "JSON array of {field,operator,value}. Operator supports: eq, neq, gt, gte, lt, lte, like",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "tahun_name",
                        "description": "sort column (single)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "asc | desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "per page, default 20",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "page, 1-based",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.TahunListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    }
                }
            }
        },
        "/master/tahun-distinct": {
            "get": {
                "description": "SELECT DISTINCT tahun_id, tahun_name FROM webapi.datacontent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master"
                ],
                "summary": "List distinct tahun from datacontent",
                "parameters": [
                    {
                        "type": "string",
                        "example": "[{\"field\":\"domain_id\",\"operator\":\"eq\",\"value\":\"0000\"}]",
                        "description": "JSON array of {field,operator,value}. Operator supports: eq, neq, gt, gte, lt, lte, like",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "tahun_id",
                        "description": "sort column (single)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "asc | desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "per page, default 20",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "page, 1-based",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.TahunDistinctListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BaseMetadata": {
            "type": "object",
            "properties": {
                "last_update_data": {
                    "type": "string"
                },
                "last_update_pipeline": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "models.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "metadata": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "services.DataContentListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DataContent"
                    }
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.BaseMetadata"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "services.DomainListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Domain"
                    }
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.BaseMetadata"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "services.SubjectListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Subject"
                    }
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.BaseMetadata"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "services.TahunDistinctListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TahunDistinct"
                    }
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.BaseMetadata"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "services.TahunListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tahun"
                    }
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.BaseMetadata"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}